+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ This project adheres to `Semantic Versioning <http://semver.org/>`__.
 [Unreleased]
 ------------
 
+- Added support for toggle command
+
 [2.0] - 2017-05-08
 ------------------